+++ memory/Ram.h
@@ -25,7 +25,7 @@ public:
 
     void store8(const uint32_t &offset, const uint8_t &value);
 
-    void store16(const uint32_t &offset, const uint8_t &value);
+    void store16(const uint32_t &offset, const uint16_t &value);
 
     uint16_t load16(const uint32_t &offset);
 };
